From uint8_t back to int

+++ Motor.ino
@@ -48,7 +48,7 @@ void Motor_Setup() {
   analogWrite(wheel_rb_pwm,0);
 }
 
-void Motor_Loop(int vx,int vy,int w,uint8_t *motor1,uint8_t *motor2,uint8_t *motor3,uint8_t *motor4){
+void Motor_Loop(int vx,int vy,int w,int *motor1,int *motor2,int *motor3,int *motor4){
   //计算每个电机分速度
   int v1, v2, v3, v4;
   v1 = (vy - vx + w * ab);
