+++ myhero.ino
@@ -1,6 +1,6 @@
 /*变量定义区*/
 int PS2_mode=1,Vx=0,Vy=0,dj1=45,dj2=45,dj3=45,Vw=0;
-uint8_t motor1 = 0, motor2 = 0, motor3 = 0, motor4 = 0;
+int motor1 = 0, motor2 = 0, motor3 = 0, motor4 = 0;
 void setup(){
   PS2_Setup();
   //Motor_Setup();
